refactor(video): extract cast and sleep timer workflows from VideoPlayerViewModel
Move cast-to-device transport glue and sleep timer fade/expiry logic into
dedicated PlayerCastController and SleepTimerController classes. This
continues the collaborator extraction pattern, reducing ViewModel complexity
and isolating engine-interaction code from UI state management.
